+++ src/filters/websocket-exceptions.filter.ts
@@ -4,11 +4,11 @@ import { BaseWsExceptionFilter, WsException } from "@nestjs/websockets";
 @Catch(HttpException)
 export class WebsocketExceptionsFilter extends BaseWsExceptionFilter {
   private readonly logger = new Logger(WebsocketExceptionsFilter.name);
-  catch(exception: HttpException, host: ArgumentsHost) {
-    this.logger.error(exception, exception.stack, {
+  catch(error: HttpException, host: ArgumentsHost) {
+    this.logger.error(error, error.stack, {
       _context: "ExceptionHandler",
     });
-    const convertedException = new WsException(exception.getResponse());
+    const convertedException = new WsException(error.getResponse());
     super.handleError(host.switchToWs().getClient(), convertedException);
     super.catch(convertedException, host);
   }